fix: Split converter into 2 classes
