+++ DUI3-DX/Converters/Rhino/Speckle.Converters.Rhino7/ToSpeckle/RhinoConverterToSpeckle.cs
@@ -1,3 +1,4 @@
+using System;
 using Speckle.Autofac.DependencyInjection;
 using Speckle.Converters.Common;
 using Speckle.Converters.Common.Objects;
